TASK: Create folder for custom components and move services grid into it (CHRT-30)
